--- src/test/java/com/briteerp/tests/AkramsPosOrderTestCase/VerifyPosmanagerperspective.java
@@ -1,53 +1,0 @@
-package com.briteerp.tests.AkramsPosOrderTestCase;
-
-import com.briteerp.pages.*;
-import com.briteerp.tests.PosOrderTestCase;
-import com.briteerp.utilities.BrowserUtils;
-import com.briteerp.utilities.Driver;
-import com.briteerp.utilities.TestBase;
-import org.openqa.selenium.support.PageFactory;
-import org.testng.Assert;
-import org.testng.annotations.BeforeMethod;
-import org.testng.annotations.Test;
-
-public class VerifyPosmanagerperspective extends TestBase {
-    public VerifyPosmanagerperspective() {
-        PageFactory.initElements(Driver.getDriver(), LoginPage.class);
-    }
-
-    @BeforeMethod(alwaysRun = true)
-    public void st1() {
-        new VerifyPosmanagerperspective();
-        new EnvironmentPage();
-        new POSHomePage();
-        new HomePage();
-        new OrderPage();
-    }
-
-    @Test
-    public void verifyTitle() throws InterruptedException {
-
-        extentLogger = report.createTest("Verify the report`s page");
-        extentLogger.info("Selecting the environment");
-        EnvironmentPage.selectEnvironment();
-        extentLogger.info("Verifying the title contain Login");
-        Assert.assertTrue(driver.getTitle().contains("Login"));
-        extentLogger.info("Entering valid user credentials and click login");
-        LoginPage.login();
-        BrowserUtils.wait(8);
-        extentLogger.info("Verifying the title contains Inbox");
-        Assert.assertTrue(driver.getTitle().contains("Inbox"));
-        extentLogger.info("Clicking the PointofSale module");
-        HomePage.pointOfSaleElement.click();
-        BrowserUtils.wait(8);
-        extentLogger.info("Verifying the title contains Point of Sale");
-        Assert.assertTrue(driver.getTitle().contains("Point of Sale"));
-        extentLogger.info("Clicking the order on the orders");
-        POSHomePage.ordersOrdersElement.click();
-        BrowserUtils.wait(8);
-        extentLogger.info("Verifying the title contains orders");
-        Assert.assertTrue(driver.getTitle().contains("Orders"));
-
-
-    }
-}
